refactor(#51) - 추가 오류사항 수정
