+++ src/main/java/com/flab/infrun/cart/application/CartsReader.java
@@ -5,7 +5,7 @@ import com.flab.infrun.cart.application.result.CartsResult.CartItemResult;
 import com.flab.infrun.cart.domain.Cart;
 import com.flab.infrun.cart.domain.CartRepository;
 import com.flab.infrun.lecture.domain.Lecture;
-import com.flab.infrun.lecture.domain.LectureRepository;
+import com.flab.infrun.lecture.domain.repository.LectureRepository;
 import com.flab.infrun.member.domain.Member;
 import com.flab.infrun.member.domain.MemberRepository;
 import java.math.BigDecimal;
